refactor: Repository 패키지 상세화
User, Article로 구분해서 관리하도록 변경
