+++ src/main/java/com/kakao/cafe/repository/article/ArticleRepository.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.article;
 
 import java.util.List;
 import java.util.Optional;
