+++ src/main/java/com/kakao/cafe/repository/article/MemoryArticleRepository.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.article;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -8,6 +8,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 
 import com.kakao.cafe.domain.article.Article;
+import com.kakao.cafe.repository.article.ArticleRepository;
 
 @Component
 public class MemoryArticleRepository implements ArticleRepository {
