+++ src/main/java/com/kakao/cafe/repository/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.user;
 
 import java.util.List;
 import java.util.Optional;
