+++ src/test/java/com/kakao/cafe/repository/ArticleRepositoryTest.java
@@ -2,7 +2,6 @@ package com.kakao.cafe.repository;
 
 import static org.assertj.core.api.Assertions.*;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -10,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import com.kakao.cafe.domain.article.Article;
+import com.kakao.cafe.repository.article.MemoryArticleRepository;
 
 class ArticleRepositoryTest {
